requirements.txt
black==24.2.0
bpy==4.0.0
pillow==10.2.0
ruff==0.2.2